pruebaoberstaff/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
pycryptodome==3.22.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
Unipath==1.1
uritemplate==4.1.1
gunicorn